+++ lab1_sequential.cpp
@@ -14,10 +14,10 @@ using namespace std;
 
 // measures distance between points x and y
 // d : dimension
-double distance(int* x, int* y, int d=DIMENSION){
+double distance(int* x, float* y, int d=DIMENSION){
 	double s = 0;
 	for(int i = 0; i < d; i++){
-		s += pow(x[i] - y[i],2);
+		s += pow(float(x[i]) - y[i],2);
 	}
 	return pow(s,0.5);
 }
@@ -36,9 +36,9 @@ std::unordered_set<int> choose_random(int N, int K, std::mt19937& gen)
 }
 
 // initializes the data_point_cluster and centroids
-void initialize(int N, int K, int * data_points, int** data_point_cluster, int** centroids){
+void initialize(int N, int K, int * data_points, int** data_point_cluster, float** centroids){
 	*data_point_cluster = (int *)malloc(sizeof(int)*N*4);
-	*centroids = (int *)malloc(sizeof(int)*K*3*MAX_ITERATIONS);
+	*centroids = (float*)malloc(sizeof(float)*K*3*MAX_ITERATIONS);
 
 	// initialize data_point_cluster
 	for(int i = 0; i < N; i++){
@@ -65,7 +65,7 @@ void initialize(int N, int K, int * data_points, int** data_point_cluster, int**
 	}
 }
 
-int assign_centroid(int K, int* x, int* centroids){
+int assign_centroid(int K, int* x, float* centroids){
 	double min_dist = MAX_DIST;
 	int cluster = 0;
 	double dist;
@@ -79,7 +79,7 @@ int assign_centroid(int K, int* x, int* centroids){
 	return cluster;
 }
 
-void kmeans_sequential(int N, int K, int* data_points, int** data_point_cluster, int** centroids, int* num_iterations){
+void kmeans_sequential(int N, int K, int* data_points, int** data_point_cluster, float** centroids, int* num_iterations){
 	int iterations = 1;
 	int points_in_cluster[K];
 	int cluster_changes = 0;
